organize forecast elements
