--- app/src/main/java/ua/sviatkuzbyt/cityweather/data/structures/ScreenState.kt
@@ -1,5 +1,0 @@
-package ua.sviatkuzbyt.cityweather.data.structures
-
-enum class ScreenState {
-    Loading, Content, Empty, Error
-}
