--- app/src/main/java/ua/sviatkuzbyt/cityweather/data/structures/forecast/ForecastScreenType.kt
@@ -1,5 +1,0 @@
-package ua.sviatkuzbyt.cityweather.data.structures.forecast
-
-enum class ForecastScreenType{
-    Today, FiveDays
-}
