+++ app/src/main/java/ua/sviatkuzbyt/cityweather/ui/elements/basic/screens/Screen.kt
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import ua.sviatkuzbyt.cityweather.data.structures.ScreenState
 import ua.sviatkuzbyt.cityweather.ui.theme.sizeSpace16
 
 @Composable
