+++ app/src/main/java/ua/sviatkuzbyt/cityweather/ui/screens/forecasts/ForecastScreen.kt
@@ -15,14 +15,15 @@ import androidx.lifecycle.viewmodel.compose.viewModel
 import ua.sviatkuzbyt.cityweather.R
 import ua.sviatkuzbyt.cityweather.data.repositories.ForecastFiveDaysRepository
 import ua.sviatkuzbyt.cityweather.data.repositories.ForecastTodayRepository
-import ua.sviatkuzbyt.cityweather.data.structures.forecast.ForecastScreenType
 import ua.sviatkuzbyt.cityweather.ui.LocalNavController
 import ua.sviatkuzbyt.cityweather.ui.elements.basic.screens.LoadScreen
 import ua.sviatkuzbyt.cityweather.ui.elements.basic.screens.Screen
 import ua.sviatkuzbyt.cityweather.ui.elements.basic.screens.TextPlug
 import ua.sviatkuzbyt.cityweather.ui.elements.basic.topbar.LabelNavigateTopBar
-import ua.sviatkuzbyt.cityweather.ui.elements.forecast.ForecastDescriptions
-import ua.sviatkuzbyt.cityweather.ui.elements.forecast.ForecastItem
+import ua.sviatkuzbyt.cityweather.ui.screens.forecasts.elements.ForecastDescriptions
+import ua.sviatkuzbyt.cityweather.ui.screens.forecasts.elements.ForecastItem
+import ua.sviatkuzbyt.cityweather.ui.screens.forecasts.elements.ForecastScreenType
+import ua.sviatkuzbyt.cityweather.ui.screens.forecasts.elements.ForecastState
 import ua.sviatkuzbyt.cityweather.ui.theme.sizeSpace16
 
 @Composable
@@ -83,9 +84,7 @@ private fun ForecastScreen(
                             }
                         }
                     }
-                    is ForecastState.Error -> TextPlug(
-                        "${stringResource(R.string.error_load)} ${state.text}"
-                    )
+                    is ForecastState.Error -> TextPlug(stringResource(state.text))
                     ForecastState.Loading -> LoadScreen()
                 }
             }
